+++ libs/ra_hal/inc/ra8d2_vreg_regs.h
@@ -226,7 +226,7 @@ typedef enum : uint16_t {
  */
 static inline volatile uint8_t* ra_vreg_dcdcctl(void)
 {
-  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + (uintptr_t)k_ra_vreg_off_dcdcctl);
+  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + k_ra_vreg_off_dcdcctl);
 }
 
 /**
@@ -235,7 +235,7 @@ static inline volatile uint8_t* ra_vreg_dcdcctl(void)
  */
 static inline volatile uint8_t* ra_vreg_vccsel(void)
 {
-  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + (uintptr_t)k_ra_vreg_off_vccsel);
+  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + k_ra_vreg_off_vccsel);
 }
 
 /**
@@ -244,7 +244,7 @@ static inline volatile uint8_t* ra_vreg_vccsel(void)
  */
 static inline volatile uint8_t* ra_vreg_lvocr(void)
 {
-  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + (uintptr_t)k_ra_vreg_off_lvocr);
+  return (volatile uint8_t*)(k_ra_vreg_sysc_base_addr + k_ra_vreg_off_lvocr);
 }
 
 #ifdef __cplusplus
